+++ src/neospy_core/src/spice/spk_segments.rs
@@ -157,7 +157,7 @@ impl SpkSegment {
             pos.into(),
             vel.into(),
             self.ref_frame,
-            self.center_id,
+            self.center_id as isize,
         ))
     }
 }
